+++ src/main/test/java/com/example/demo/service/impl/TransferServiceTest.java
@@ -8,6 +8,7 @@ import junit.framework.TestCase;
 import org.junit.Before;
 import org.junit.Test;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -33,7 +34,7 @@ public class TransferServiceTest extends TestCase {
         transfer.setIdFromBill(1);
         transfer.setIdToUser(2);
         transfer.setIdToBill(2);
-        transfer.setSumTransaction(500);
+        transfer.setSumTransaction(BigDecimal.valueOf(500));
         transferStorage.add(transfer);
         verify(transferStorage).add(transfer);
     }
@@ -80,11 +81,11 @@ public class TransferServiceTest extends TestCase {
     @Test
     public void test_sumBalanceTransaction_Ok() {
         Bill bill = new Bill();
-        bill.setBalance(6);
+        bill.setBalance(BigDecimal.valueOf(6));
         bill.setId(2);
-        int sumDigit = 3;
+        BigDecimal sumDigit = BigDecimal.valueOf(3);
         when(billStorage.findBillFromId(6)).thenReturn(bill);
-        Bill returnBill = subj.sumBalanceTransaction(6, 3);
+        Bill returnBill = subj.sumBalanceTransaction(6, BigDecimal.valueOf(3));
         verify(billStorage).updateBill(returnBill);
         assertEquals(bill.getBalance(), returnBill.getBalance());
     }
@@ -92,9 +93,9 @@ public class TransferServiceTest extends TestCase {
     @Test
     public void test_reduceBalance_Ok() {
         Bill bill = new Bill();
-        bill.setBalance(9);
+        bill.setBalance(BigDecimal.valueOf(9));
         bill.setId(2);
-        int reduceBalance = 2;
+        BigDecimal reduceBalance = BigDecimal.valueOf(2);
         when(billStorage.findBillFromId(2)).thenReturn(bill);
         Bill returnBill = subj.reduceBalance(2, reduceBalance);
         verify(billStorage).updateBill(returnBill);
